feat(ui): move MCP server Test Connection into details drawer
