+++ client/src/components/servers/MCPServerDetailsPanel.test.tsx
@@ -124,6 +124,18 @@ describe("MCPServerDetailsPanel", () => {
     expect(region).toHaveAttribute("aria-hidden", "false");
   });
 
+  it("shows the Test Connection tab with the server URL prefilled", async () => {
+    const user = userEvent.setup();
+    renderWithProviders(
+      <MCPServerDetailsPanel server={mockServer} error={null} open={true} onClose={() => {}} />,
+    );
+
+    await user.click(screen.getByRole("tab", { name: /test connection/i }));
+    expect(screen.getByDisplayValue("http://test.example.com")).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: /^test connection$/i })).toBeInTheDocument();
+    expect(screen.queryByRole("button", { name: /search components/i })).not.toBeInTheDocument();
+  });
+
   it("displays loading state while fetching components", () => {
     // Suspend all three fetches so componentsLoading stays true.
     server.use(
